# go-service/internal/docs/swagger.yaml
definitions:
  dto.CreateProductRequest:
    properties:
      name:
        example: Premium Widget
        type: string
      price:
        example: 29.99
        type: number
      tags:
        example:
        - '["gadget"'
        - '"home"]'
        items:
          type: string
        type: array
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: resource not found
        type: string
    type: object
  dto.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  dto.ProductResponse:
    properties:
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      id:
        example: a1b2c3
        type: string
      name:
        example: Premium Widget
        type: string
      price:
        example: 29.99
        type: number
      tags:
        example:
        - '["gadget"'
        - '"home"]'
        items:
          type: string
        type: array
      updated_at:
        example: "2024-01-02T15:04:05Z"
        type: string
    type: object
  dto.UpdateProductRequest:
    properties:
      name:
        example: Premium Widget
        type: string
      price:
        example: 29.99
        type: number
      tags:
        example:
        - '["gadget"'
        - '"home"]'
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Returns the health status of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Health check
      tags:
      - Health
  /products:
    get:
      description: Returns all stored products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Creates a new product from the provided payload
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Removes a product permanently
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete product
      tags:
      - Products
    get:
      description: Fetch a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update product
      tags:
      - Products
swagger: "2.0"